Only rejoin channel after socket has reconnected
Channels listen to a `broadcast` that gives the `Connectivity` of the `socket` and wait for the socket to reconnect when disconnected before even starting their rejoin timers.  Reconnect and rejoin are now multipliers of the original connect and join times instead of a fixed number of seconds.  This means the reconnect and rejoin will scale with the speed of the server.  The first reconnect and rejoin has a timeout of 0 seconds to allow a cycle through the state machines, but nothing else to allow for immediate reconnect and rejoin if possible.